+++ weblate/checks/tests/test_markup_checks.py
@@ -340,3 +340,37 @@ class SafeHTMLCheckTest(CheckTestCase):
         self.test_failure_1 = (safe, '<a href="javascript:foo()">link</a>', "safe-html")
         self.test_failure_2 = (safe, '<a href="#" onclick="x()">link</a>', "safe-html")
         self.test_failure_3 = (safe, '<iframe src="xxx"></iframe>', "safe-html")
+
+    def test_markdown(self):
+        self.do_test(
+            False,
+            (
+                "See <https://weblate.org>",
+                "Viz <https://weblate.org>",
+                "md-text,safe-html",
+            ),
+        )
+        self.do_test(
+            True,
+            (
+                "See <https://weblate.org>",
+                "Viz <https://weblate.org>",
+                "safe-html",
+            ),
+        )
+        self.do_test(
+            False,
+            (
+                "See <noreply@weblate.org>",
+                "Viz <noreply@weblate.org>",
+                "md-text,safe-html",
+            ),
+        )
+        self.do_test(
+            True,
+            (
+                "See <noreply@weblate.org>",
+                "Viz <noreply@weblate.org>",
+                "safe-html",
+            ),
+        )